Some refactoring.
