--- library/src/main/java/com/aminography/primedatepicker/fragment/dataholder/PickedDayEmptyDataHolder.kt
@@ -1,5 +1,0 @@
-package com.aminography.primedatepicker.fragment.dataholder
-
-data class PickedDayEmptyDataHolder(
-        override val id: String
-) : BasePickedDayDataHolder(id)
