+++ library/src/main/java/com/aminography/primedatepicker/fragment/header/dataholder/BasePickedDayDataHolder.kt
@@ -1,4 +1,4 @@
-package com.aminography.primedatepicker.fragment.dataholder
+package com.aminography.primedatepicker.fragment.header.dataholder
 
 abstract class BasePickedDayDataHolder(
         open val id: String
