+++ library/src/main/java/com/aminography/primedatepicker/fragment/header/dataholder/PickedDayEmptyDataHolder.kt
@@ -1,0 +1,5 @@
+package com.aminography.primedatepicker.fragment.header.dataholder
+
+data class PickedDayEmptyDataHolder(
+        override val id: String = ""
+) : BasePickedDayDataHolder(id)
