+++ library/src/main/java/com/aminography/primedatepicker/fragment/header/viewholder/PickedDayViewHolder.kt
@@ -1,12 +1,12 @@
-package com.aminography.primedatepicker.fragment.viewholder
+package com.aminography.primedatepicker.fragment.header.viewholder
 
 import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.Typeface
 import android.view.LayoutInflater
 import com.aminography.primedatepicker.R
-import com.aminography.primedatepicker.fragment.adapter.BaseAdapter
-import com.aminography.primedatepicker.fragment.dataholder.PickedDayDataHolder
+import com.aminography.primedatepicker.fragment.header.adapter.BaseAdapter
+import com.aminography.primedatepicker.fragment.header.dataholder.PickedDayDataHolder
 import kotlinx.android.synthetic.main.list_item_picked_day.view.*
 import java.util.*
 
@@ -23,8 +23,6 @@ class PickedDayViewHolder(
         with(itemView) {
             typeface?.run {
                 twoLineTextView.typeface = this
-                twoLineTextView.firstLabelTextSize = 90
-                twoLineTextView.secondLabelTextSize = 40
             }
         }
     }
